requirements.txt
Pillow
imutils
selenium
Appium-Python-Client
pytest
pytesseract
opencv-python
Click
pyaml
requests
black
isort
-e .